+++ CHANGELOG.md
@@ -40,6 +40,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - SB_WEB_FORCE_ASSET_VERSION Int to forced update of external CSS and JavaScript files
 - SB_APP_ROOT_ABSOLUTE_URL The absolute URL of the root of the application
 - show SQL statements in Tracy
+- database load checked, if fails an Exception is thrown
+- if Location redirection fails, a nice redirection HTML page ( redirection.latte ) is displayed
 
 ### Changed
 - APP_COLLECTION -> APP_MAPPING